Refactoring mirror > sync + activitypub
