+++ src/middleware/packages/activitypub/services/activitypub/subservices/actor.js
@@ -1,8 +1,8 @@
 const fetch = require('node-fetch');
 const { namedNode, literal, triple, variable } = require('@rdfjs/data-model');
 const { MIME_TYPES } = require('@semapps/mime-types');
-const { ACTOR_TYPES, AS_PREFIX } = require('../constants');
-const { delay, defaultToArray, getSlugFromUri } = require('../utils');
+const { ACTOR_TYPES, AS_PREFIX } = require('../../../constants');
+const { delay, defaultToArray, getSlugFromUri } = require('../../../utils');
 
 const ActorService = {
   name: 'activitypub.actor',
