+++ src/middleware/packages/activitypub/services/activitypub/subservices/inbox.js
@@ -1,7 +1,7 @@
 const { MIME_TYPES } = require('@semapps/mime-types');
-const { objectIdToCurrent, collectionPermissionsWithAnonRead } = require('../utils');
-const ControlledCollectionMixin = require('../mixins/controlled-collection');
-const { ACTOR_TYPES } = require('../constants');
+const { objectIdToCurrent, collectionPermissionsWithAnonRead } = require('../../../utils');
+const ControlledCollectionMixin = require('../../../mixins/controlled-collection');
+const { ACTOR_TYPES } = require('../../../constants');
 
 const InboxService = {
   name: 'activitypub.inbox',
