+++ src/middleware/packages/activitypub/services/activitypub/subservices/registry.js
@@ -1,9 +1,9 @@
 const urlJoin = require('url-join');
 const pathJoin = require('path').join;
 const { quad, namedNode } = require('@rdfjs/data-model');
-const getCollectionRoute = require('../routes/getCollectionRoute');
-const { defaultToArray, getSlugFromUri } = require('../utils');
-const { ACTOR_TYPES, FULL_ACTOR_TYPES, AS_PREFIX} = require('../constants');
+const getCollectionRoute = require('../../../routes/getCollectionRoute');
+const { defaultToArray, getSlugFromUri } = require('../../../utils');
+const { ACTOR_TYPES, FULL_ACTOR_TYPES, AS_PREFIX} = require('../../../constants');
 
 const RegistryService = {
   name: 'activitypub.registry',
