--- src/middleware/packages/mirror/README.md
@@ -1,5 +1,0 @@
-# @semapps/mirror
-
-Mirror module for SemApps
-
-[Documentation](https://semapps.org/docs/middleware/mirror/index)
