--- src/middleware/packages/mirror/bots-container.json
@@ -1,6 +1,0 @@
-{
-  "path": "/bots",
-  "acceptedTypes": ["Application"],
-  "dereference": ["sec:publicKey"],
-  "readOnly": true
-}
