--- src/middleware/packages/mirror/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  MirrorService: require('./service'),
-  botsContainer: require('./bots-container.json')
-};
