+++ src/middleware/packages/sync/README.md
@@ -1,0 +1,5 @@
+# @semapps/sync
+
+Synchronization tools for SemApps
+
+[Documentation](https://semapps.org/docs/middleware/sync/index)
