+++ src/middleware/packages/sync/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  ObjectsWatcherMiddleware: require('./middlewares/objects-watcher'),
+  AggregatorService: require('./services/aggregator'),
+  MirrorService: require('./services/mirror'),
+  SynchronizerService: require('./services/synchronizer')
+};
