+++ src/middleware/tests/interop/initialize.js
@@ -1,11 +1,11 @@
 const fse = require('fs-extra');
 const path = require('path');
 const { ServiceBroker } = require('moleculer');
-const { ACTOR_TYPES, SynchronizerService } = require('@semapps/activitypub');
+const { ACTOR_TYPES, RelayService } = require('@semapps/activitypub');
 const { AuthLocalService } = require('@semapps/auth');
 const { CoreService, defaultOntologies } = require('@semapps/core');
 const { InferenceService } = require('@semapps/inference');
-const { ObjectsWatcherMiddleware } = require("@semapps/activitypub");
+const { MirrorService, ObjectsWatcherMiddleware } = require('@semapps/sync');
 const { WebAclMiddleware } = require('@semapps/webacl');
 const { WebIdService } = require('@semapps/webid');
 const CONFIG = require('../config');
@@ -75,6 +75,16 @@ const initialize = async (port, mainDataset, accountsDataset, serverToMirror) =>
     }
   });
 
+  await broker.createService(RelayService);
+
+  if (serverToMirror) {
+    await broker.createService(MirrorService, {
+      settings: {
+       servers: [serverToMirror]
+      }
+    });
+  }
+
   await broker.createService(AuthLocalService, {
     settings: {
       baseUrl,
@@ -92,12 +102,11 @@ const initialize = async (port, mainDataset, accountsDataset, serverToMirror) =>
   await broker.createService(InferenceService, {
     settings: {
       baseUrl,
-      offerRemoteServers: true,
+      acceptFromRemoteServers: true,
+      offerToRemoteServers: true,
       ontologies: defaultOntologies
     }
   });
-
-  await broker.createService(SynchronizerService);
 
   await broker.start();
